app: telebot: restructure files
